+++ scripts/2014/feats/elemental-adept-fire.js
@@ -49,7 +49,7 @@ try {
 	}
 	else if (lastArg.macroPass === "postDamageRoll") {
 		// check pre-requisites
-		// make sure the dmage type matches
+		// make sure the damage type matches
 		if (!theItem.system.damage.parts.map(i=>i[1]).includes(damageType)) {
 			return;
 		}