+++ scripts/macros/add-heroic-inspiration.js
@@ -1,0 +1,4 @@
+for (let tok of canvas.tokens.controlled) {
+    await tok.actor.update({'system.attributes.inspiration' : true});
+}
+