+++ scripts/macros/apply-effect.js
@@ -1,0 +1,3 @@
+for (let tok of canvas.tokens.controlled) {
+    await HomebrewEffects.applyGrappledEffect(tok.actor, this.uuid, 13);
+}