+++ scripts/macros/close-all-doors.js
@@ -1,0 +1,1 @@
+canvas.walls.updateAll({ds: 0}, w => w.document.ds===1)