+++ scripts/macros/remove-all-effects.js
@@ -1,0 +1,5 @@
+for (let tok of canvas.tokens.controlled) {
+    let removeList = tok.actor?.effects.map(e=>e.id);
+    if(removeList)
+        await MidiQOL.socket().executeAsGM("removeEffects", { actorUuid: tok.actor.uuid, effects: removeList });
+}