+++ changelog.md
@@ -3,3 +3,12 @@
 * Restructuring and organization of the homebrew compendiums
 * All core backgrounds updated in Homebrew
 * Brought in all core features (classes, subclasses, races) that need to be automated for a baseline set of features
+
+# 12.4.1
+* All races updated
+* All origin feats updated
+* Most general feats updated
+* All fighting style feats updated
+* All Tasha's feats updated
+* Added Air Genasi and Kobold for RotD characters
+* All classes updated